102-fibonacci.c fix

+++ functions_nested_loops/102-fibonacci.c
@@ -7,14 +7,14 @@
  */
 void fibonacci(void)
 {
-	int nb1, nb2 = 1;
+	long nb1, nb2 = 1;
 	int i, temp;
 
 	for (i = 0; i < 50; i++)
 	{
 		temp = nb1;
 		nb1 += nb2;
-		printf("%d", nb1);
+		printf("%ld", nb1);
 		nb2 = temp;
 		if (i < 49)
 			printf(", ");
